% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclid_Distance.R
\name{Euclid_Distance}
\alias{Euclid_Distance}
\title{Euclid_Distance function}
\usage{
Euclid_Distance(X, v)
}
\arguments{
\item{X}{n x m matrix}

\item{v}{1 x m vector}
}
\value{
(Euclidean) distance between the rows of X and the vector v
}
\description{
Euclid_Distance function
}
\examples{
X = matrix(c(17,40,20,39,30.1,55.5), nrow = 2)
v = c(10.2,30,45)
Euclid_Distance(X,v)
}
