% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN_Euc.R
\name{KNN_Euc}
\alias{KNN_Euc}
\title{KNN_Euc function}
\usage{
KNN_Euc(
  X_test,
  X_pred,
  Y_test,
  Y_pred,
  K,
  method = c("Euc", "Man", "Min", "Cos"),
  pred_weights = FALSE,
  p = NULL
)
}
\arguments{
\item{X_test}{n x k length numeric dataframe or matrix}

\item{X_pred}{r x k length numeric dataframe or matrix}

\item{Y_test}{n x 1 length dataframe or matrix  label column which is the label column for the X_test dataframe}

\item{Y_pred}{r x 1 length dataframe or matrix label column which is the label column for the X_pred dataframe}

\item{K}{integer value; the number of neighbors}

\item{method}{Determines how distance will be calculated for nearest-neighbors. Choose from the following:
"Euc": will calculate using the Euclidean Distance
"Man": will calculate using the Manhattan Distance
"Min": will calculate using the Minkowski Distance
"Cos": will calculate using the Cosine-Similarity Distance}

\item{pred_weights}{default is FALSE; if TRUE will calculate weights for predictions of K-nearest Neighbors
by calculating the inverse of the distances of the K-nearest neighbors if Y_test is of numeric type.}
}
\value{
r x 1 vector of predicted labels for the X_pred dataframe
}
\description{
KNN_Euc function
}
