% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhattan_Distance.R
\name{Manhattan_Distance}
\alias{Manhattan_Distance}
\title{Manhattan_Distance function}
\usage{
Manhattan_Distance(X, v)
}
\arguments{
\item{X}{n x p matrix}

\item{v}{1 x p vector}
}
\value{
column vector of (Manhattan) Distances calculated between the rows of X and vector v
}
\description{
Manhattan_Distance function
}
