% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nearest_Neighbors.R
\name{Nearest_Neighbors}
\alias{Nearest_Neighbors}
\title{Nearest_Neighbors function}
\usage{
Nearest_Neighbors(X, observation, K, method, p = NULL)
}
\arguments{
\item{X}{n x m matrix}

\item{observation}{1 x m vector which is a row of the matrix X}

\item{K}{positive integer that is the number of nearest-neighbors}

\item{method}{Which distance equation will be used to calculate the nearest-neighbors.}

\item{p}{If supplied, the positive integer used in calculating the Minkowski distance}
}
\value{
A list with first item is the 1 x K vector of indices of the nearest-neighbors and
the second item is the 1 x K vector of distances to the nearest-neighbors.
}
\description{
Nearest_Neighbors function
}
