% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_NN.R
\name{Prediction_NN}
\alias{Prediction_NN}
\title{Prediction_NN function}
\usage{
Prediction_NN(X, Y, weights = NULL)
}
\arguments{
\item{X}{n x m matrix}

\item{Y}{1 x m vector which is a single row of the matrix X}

\item{weights}{If supplied will be given the vector of distances calculated from the Nearest_Neighbors function}
}
\value{
vector of predicted labels
}
\description{
Prediction_NN function
}
