% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cos_Distance.R
\name{Cos_Distance}
\alias{Cos_Distance}
\title{Cosine_Distance function}
\usage{
Cos_Distance(X, v)
}
\arguments{
\item{X}{n x k matrix}

\item{v}{1 x k vector}
}
\value{
vector of (Cosine Similarity) Distances calculated between the rows of X and the vector v
}
\description{
Cosine_Distance function
}
