% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cos_Distance.R
\name{Cos_Distance}
\alias{Cos_Distance}
\title{Cos_Distance function}
\usage{
Cos_Distance(X, v)
}
\arguments{
\item{X}{n x m matrix}

\item{v}{1 x m vector}
}
\value{
column matrix of the (Cosine Similarity) Distances calculated between the rows of X and the vector v
}
\description{
Cos_Distance function
}
\examples{
X = matrix(c(1,1,0,1,1,0), nrow = 2, byrow = FALSE)
v = c(0,0,1)
Cos_Distance(X,v)
# Returns the following:
# [0.2928932, 1.0000000]'

}
