% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhattan_Distance.R
\name{Manhattan_Distance}
\alias{Manhattan_Distance}
\title{Manhattan_Distance function}
\usage{
Manhattan_Distance(X, v)
}
\arguments{
\item{X}{n x m matrix}

\item{v}{1 x m vector}
}
\value{
column matrix of (Manhattan) Distances calculated between the rows of X and vector v
}
\description{
Manhattan_Distance function
}
\examples{
X = matrix(c(1,17,19,21,1,0), nrow = 2)
v = c(10,15,3)
Manhattan_Distance(X,v)
# returns the following:
# [15,16]'
}
