% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minkowski_Distance.R
\name{Minkowski_distance}
\alias{Minkowski_distance}
\title{Minkowski Distance function}
\usage{
Minkowski_distance(X, v, p)
}
\arguments{
\item{X}{n x k matrix or numeric dataframe}

\item{v}{1 x k vector}

\item{p}{positive non-zero integer}
}
\value{
vector of (Minkowski) distances between the rows of X and a vector v
}
\description{
Minkowski Distance function
}
